Fix some issues bilding on AIX with gcc.
AIX has libiconv, but the cmake code to find it doesn't find it, so simply add -liconv to the link.
If TRUE/FALSE have been defined, undefine them to prevent compilation errors.
In qglobal.h, include inittypes so that we don't have an include order problem where qglobal.h defines different typedefs depending on if inttypes.h has been included or not.
Provide a definition for PRIu64 if it wasn't defined by the system.